fixed some issues with slots and NIO buffers not being released
removed the thread pool from PlayerHandler.java, its overkill and also just adds a layer of things to go wrong. There was an issue when a player would login with the same account, it would give that slot to that player even tho it never completed the login that slot would be taken up by that player. it now checks to make sure login is complete