anti ddos/packet flood (#44)
* NIO networking from Java Socket IO. Needs some testing but should be working.

* "lag" mostly with  private static final long SELECTOR_TIMEOUT = 10ms now

* fixed some issues with slots and NIO buffers not being released

removed the thread pool from PlayerHandler.java, its overkill and also just adds a layer of things to go wrong. There was an issue when a player would login with the same account, it would give that slot to that player even tho it never completed the login that slot would be taken up by that player. it now checks to make sure login is complete

* debugging/error with numbering slots

fixed

* slot issue

issue where if two players both login at the same they both get ths same slot. FIXED. TODO logout method has either a race condition or timing issue. have to click it a lot to logout otherwise the socket handler kills the connection it seems before logout takes place

* fixing stuff

* fixed logging out and resource clean up

* Update PlayerHandler.java

* Update PlayerHandler.java

reuse slots if connection is not completed/null

* Update Client.java to Fix null pointer

* Update Memory.java

need more info on what is in the arrays

* Update ServerConnectionHandler.java

anti-DDOS

* Update PacketParser.java

limit packets per second. This will need tested. Also not sure the best way to handle it. maybe just log out the user? or delay packets?

* Update Client.java