Improves client-server communication efficiency
Implements caching for UI text, walkable interfaces, and player context menus to reduce redundant data transmission.

Avoids unnecessary packet sends by checking for changes before sending updates, improving performance.

This may not seem as big of a deal but the server was sending a lot of strings/ui updates per tick per player and it was actually a big overhead